+++ Sources/API/CustomCargoAPI.h
@@ -20,7 +20,7 @@
 // =======================================================================================
 
 #pragma once
-#include "UCSO_CustomCargo.h"
+#include "CustomCargo.h"
 
 namespace UCSO
 {