+++ Sources/Cargo/Cargo.h
@@ -22,7 +22,7 @@
 #pragma once
 #include "..\API\Helper.h"
 
-DLLCLBK const char* GetUCSOVersion() { return _strdup("1.0.1"); }
+DLLCLBK const char* GetUCSOVersion() { return _strdup("1.1"); }
 
 namespace UCSO
 {
@@ -65,15 +65,10 @@ namespace UCSO
 			MANUAL
 		};
 
-		static bool configLoaded;
-		static double containerMass;
-		static bool enableFocus;
-		static bool drainUnpackedResources;
-
 		DataStruct dataStruct;
 
-		std::string packedMesh = "UCSO\\";
-		std::string unpackedMesh = "UCSO\\";
+		std::string packedMesh = "UCSO/";
+		std::string unpackedMesh = "UCSO/";
 
 		double resourceContainerMass = 0;
 		double unpackedSize;
@@ -96,3 +91,8 @@ namespace UCSO
 		void ThrowWarning(const char* warning);
 	};
 }
+
+bool configLoaded = false;
+double containerMass = 85;
+bool enableFocus = false;
+bool drainUnpackedResources = false;